html/browsers/browsing-the-web/navigating-across-documents/006.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate">
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation">
<title>Link with onclick form submit is aborted by href navigation </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe name="test"></iframe>
<form target="test" action="click.html"></form>
<a target="test" onclick="document.forms[0].submit()" href="href.html">Test</a>
<script>
var test = async_test();

test.step(function() {
  document.querySelector("a").click();
});

window.onmessage = test.step_func(function(e) {
  assert_equals(e.data, "href");
  test.done();
});
</script>

<!--
1. This test creates an iframe, a form with `action="click.html", and
    an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will attempt to submit the form, which targets the iframe.
3. Navigating the iframe to the anchor's `href` will abort the not-yet-mature iframe navigation.
4. If the semantics are implemented correctly, a message "href"
    from "href.html" will be received.

-->
